+++ src/app.module.ts
@@ -3,15 +3,15 @@ import { APP_FILTER } from '@nestjs/core';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { AuthModule } from './auth/auth.module';
-import CatchException from './common/exceptions/http-exception.filter';
+import CatchException from './common/filters/http-exception.filter';
 import { ConfigModule } from './config/config.module';
 import { ConfigService } from './config/config.service';
+import { FacilityModule } from './facility/facility.module';
 import { FloorModule } from './floor/floor.module';
 import { ReservationModule } from './reservation/reservation.module';
-import { SeatModule } from './seat/seat.module';
 import { RoomModule } from './room/room.module';
+import { SeatModule } from './seat/seat.module';
 import { UserModule } from './user/user.module';
-import { FacilityModule } from './facility/facility.module';
 
 @Module({
   imports: [
@@ -29,11 +29,6 @@ import { FacilityModule } from './facility/facility.module';
     FacilityModule,
   ],
   controllers: [],
-  providers: [
-    {
-      provide: APP_FILTER,
-      useClass: CatchException,
-    },
-  ],
+  providers: [],
 })
 export class AppModule {}
