+++ src/main.ts
@@ -1,9 +1,14 @@
 import * as morgan from 'morgan';
 
-import { ValidationPipe } from '@nestjs/common';
+import {
+  BadRequestException,
+  ValidationError,
+  ValidationPipe,
+} from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 
 import { AppModule } from './app.module';
+import CatchException from './common/filters/http-exception.filter';
 import { LoggerService } from './common/utils/logger/logger.service';
 import { ConfigModule } from './config/config.module';
 import { ConfigService } from './config/config.service';
@@ -18,8 +23,21 @@ async function bootstrap() {
       whitelist: true,
       forbidNonWhitelisted: true,
       transform: true,
+      exceptionFactory: (errors: ValidationError[]) => {
+        const messages = errors.map(error => {
+          return {
+            error: `${error.property} has wrong value ${error.value}.`,
+            message: Object.values(error.constraints).join(''),
+          };
+        });
+
+        return new BadRequestException({ validation: messages });
+      },
     }),
   );
+
+  // Filter
+  app.useGlobalFilters(new CatchException());
 
   // Logger
   const loggerService = app.select(ConfigModule).get(LoggerService);
