+++ src/reservation/reservation.dto.ts
@@ -1,4 +1,4 @@
-import { IsArray, IsDate, IsNumber } from 'class-validator';
+import { IsArray, IsDate, IsNumber, IsOptional } from 'class-validator';
 
 export class CreateRoomReservationDto {
   @IsDate()
@@ -11,9 +11,11 @@ export class CreateRoomReservationDto {
   status: number;
 
   @IsNumber()
+  @IsOptional()
   seatId?: number;
 
   @IsNumber()
+  @IsOptional()
   roomId?: number;
 
   @IsNumber()
@@ -33,9 +35,11 @@ export class CreateSeatReservationDto {
   status: number;
 
   @IsNumber()
+  @IsOptional()
   seatId?: number;
 
   @IsNumber()
+  @IsOptional()
   roomId?: number;
 
   @IsNumber()
